pyproject.toml
[project]
name = "espn-fantasy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

[dependency-groups]
dev = [
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "espn-api==0.44.1",
    "fonttools==4.57.0",
    "idna==3.10",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.1",
    "numpy==2.2.5",
    "packaging==25.0",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "requests==2.32.3",
    "six==1.17.0",
    "streamlit==1.22.0",
    "setuptools==79.0.0",
    "tzdata==2025.2",
    "urllib3==2.2.3",
]
